Fix dc removal bug, rename one pole.
Fix an issue where the  DC-removal step wasn't functioning properly and
was in fact a DC-preserving step.

This was caused partially by retaining the name "OnePoleLowPassFilter"
when the filter is in fact a low or high pass filter depending on
configuration. As such, rename the structure to "OnePoleFilter" so that
future users don't make the same mistake.

Also, use the handy "default" attribute for implementing a default value
for the filter type enum.